sharepoint/requirements.txt
langchain
langchain-community
langchain-openai
langchain-google-vertexai
google-cloud-aiplatform
google-cloud-discoveryengine
langchain-google-community


O365
chromadb
langchain-chroma
streamlit
python-dotenv
tiktoken
unstructured
networkx
pypdf

